Further fixes for API explorer in App Engine Java 11
